feat: 부스 공지 및 등록
